EngineerAPI/requirements.txt
discord.py
dnspython
psycopg2-binary
python-dotenv